+++ widgetbook/lib/main.directories.g.dart
@@ -1,0 +1,27 @@
+// dart format width=80
+// coverage:ignore-file
+// ignore_for_file: type=lint
+// ignore_for_file: unused_import, prefer_relative_imports, directives_ordering
+
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// **************************************************************************
+// AppGenerator
+// **************************************************************************
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'package:flame_fuse_widgetbook/stories/fuses/core_story.dart'
+    as _flame_fuse_widgetbook_stories_fuses_core_story;
+import 'package:widgetbook/widgetbook.dart' as _widgetbook;
+
+final directories = <_widgetbook.WidgetbookNode>[
+  _widgetbook.WidgetbookComponent(
+    name: 'Fuse',
+    useCases: [
+      _widgetbook.WidgetbookUseCase(
+        name: 'Default',
+        builder: _flame_fuse_widgetbook_stories_fuses_core_story.buildFuseStory,
+      ),
+    ],
+  ),
+];